Inline the VM report config form footer
